python code and bindings for faster xml parsing.

+++ R/helpers.R
@@ -186,22 +186,35 @@ read_csv(path) %>%
 
 #' Download xml xtab data from url, tag with code for caching
 #' useful for older (2006) data that does not come as csv option
+#'
+#' @param code The statcan table code, needed for caching and for locating the table to extract
+#' @param url The url to the xml table download. Only needed if the table is not cached, but
+#' reommended to always include for reproducibility
+#' @param refresh Will refresh cached data if set to TRUE, default is FALSE
+#' @param time_value If set, will use this as the time variable, otherwise extract the time variable from the data.
+#' Setting this will slightly speed up the parsing.
+#' @param temp A path to the downloaded xml zip file. Useful if file has already been downloaded and should not be
+#' downloaded again.
+#'
 #' @export
-xml_xtab_for <- function(code,url,refresh=FALSE){
+xml_xtab_for <- function(code,url,refresh=FALSE,time_value=NA,temp=NA){
   data <- NA
   path <- file.path(getOption("custom_data_path"),paste0(code,".rda"))
   if (!file.exists(path) | refresh) {
-    temp <- tempfile()
     exdir <- tempdir()
-    download.file(url,temp)
-    utils::unzip(temp,exdir=exdir)
-    unlink(temp)
+    if (is.na(temp)) {
+      temp <- tempfile()
+      download.file(url,temp)
+      utils::unzip(temp,exdir=exdir)
+      unlink(temp)
+    } else {
+      utils::unzip(temp,exdir=exdir)
+    }
     # read xml xtab for 2006 data.
-    library(xml2)
     message("Parsing XML")
     xml_structure <- read_xml(file.path(exdir,paste0("Structure_",code,".xml")))
     xml_data <- read_xml(file.path(exdir,paste0("Generic_",code,".xml")))
-    unlink(exdir)
+    unlink(exdir,recursive = TRUE)
 
     str_concepts <- xml_structure %>% xml_find_all("//structure:ConceptScheme")
     var_concepts <- str_concepts %>%
@@ -213,13 +226,15 @@ xml_xtab_for <- function(code,url,refresh=FALSE){
       xml_structure %>% xml_find_all(paste0("//structure:ConceptScheme/structure:Concept[@id='",c,"']/structure:Name[@xml:lang='en']")) %>%
         xml_text}) %>% unlist
 
-    concept_lookup <- set_names(concept_names,concepts)
+    concept_lookup <- rlang::set_names(concept_names,concepts)
 
     descriptions_for_code <- function(c){
+      c <- gsub("0$","",c)
       base <- xml_structure %>% xml_find_all(paste0("//structure:CodeList[@id='CL_",toupper(c),"']/structure:Code"))
-      set_names(
-        base %>% map(function(e){e %>% xml_find_all(".//structure:Description[@xml:lang='en']") %>% xml_text}) %>% unlist %>% trimws(),
-        base %>% map(function(e){xml_attr(e,"value")}) %>% unlist
+      desc_text = ifelse(length(xml_find_all(base[1] ,".//structure:Description[@xml:lang='en']"))==0,".//structure:Description",".//structure:Description[@xml:lang='en']")
+      rlang::set_names(
+        base %>% purrr::map(function(e){e %>% xml_find_all(desc_text) %>% xml_text}) %>% unlist %>% trimws(),
+        base %>% purrr::map(function(e){xml_attr(e,"value")}) %>% unlist
       )
     }
 
@@ -230,30 +245,37 @@ xml_xtab_for <- function(code,url,refresh=FALSE){
     #l <- lapply(concepts,function(c){series %>% xml_find_all(paste0(".//generic:Value[@concept='",c,"']")) %>% xml_attr("value")})
 
     time_data <- function(series){
-      message("Extracting year data")
-      series %>% xml_find_all(".//generic:Time") %>% xml_text
+      if (!is.na(time_value)) {
+        result=rep(time_value,length(series))
+      } else {
+        message("Extracting year data")
+        series %>% xml_find_all(".//generic:Time") %>% xml_text
+      }
     }
     value_data <- function(series){
       message("Extracting values data")
-      series %>% xml_find_all(".//generic:ObsValue") %>% xml_attr("value")
+      #series %>% xml_find_all(".//generic:ObsValue") %>% xml_attr("value")
+      series %>% xml_find_all("./generic:Obs") %>% xml_find_first("./generic:ObsValue") %>% xml_attr("value")
     }
     code_data <- function(series,c){
       message(paste0("Extracting data for ",concept_lookup[c]))
       series %>% xml_find_all(paste0(".//generic:Value[@concept='",c,"']")) %>% xml_attr("value")
     }
 
-
     df=concepts %>%
-      map(function(c){code_data(series,c)}) %>%
-      set_names(concept_names) %>%
-      as.tibble() %>%
-      bind_cols(tibble(Year = time_data(series), Value = value_data(series)))
+      purrr::map(function(c){code_data(series,c)}) %>%
+      rlang::set_names(concept_names) %>%
+      tibble::as.tibble() %>%
+      dplyr::bind_cols(tibble::tibble(Year = time_data(series), Value = value_data(series)))
 
-    for (i in seq(2,length(concepts),1)) {
+    for (i in seq(1,length(concepts),1)) {
       c=concepts[i]
       n=concept_names[i] %>% as.name
       lookup <- descriptions_for_code(c)
-      df <- df %>% mutate(!!n := lookup[!!n])
+      nid=paste0(n," ID") %>% as.name
+      df <- df %>%
+        mutate(!!nid := !!n) %>%
+        mutate(!!n := lookup[!!nid])
     }
 
     fix_ct_geo_format <- function(geo){
@@ -261,7 +283,7 @@ xml_xtab_for <- function(code,url,refresh=FALSE){
     }
 
     data <- df  %>%
-      rename(GeoUID=Geography) %>%
+      rename(GeoUID=`Geography ID`) %>%
       mutate(GeoUID = fix_ct_geo_format(GeoUID),
              Value=as.numeric(Value))
 
@@ -272,17 +294,168 @@ xml_xtab_for <- function(code,url,refresh=FALSE){
   data
 }
 
+#' Download xml xtab data from url, tag with code for caching
+#' useful for (2001) census profile data
+#' @export
+xml_census_2001_profile <- function(code,url,refresh=FALSE,time_value=NA,temp=NA){
+  data <- NA
+  path <- file.path(getOption("custom_data_path"),paste0(code,".rda"))
+  if (!file.exists(path) | refresh) {
+    exdir <- tempdir()
+    if (is.na(temp)) {
+      temp <- tempfile()
+      download.file(url,temp)
+      utils::unzip(temp,exdir=exdir)
+      unlink(temp)
+    } else {
+      utils::unzip(temp,exdir=exdir)
+    }
+    # read xml xtab for 2006 data.
+    message("Parsing XML")
+    xml_structure <- read_xml(file.path(exdir,paste0("Structure_",code,".xml")))
+    xml_data <- read_xml(file.path(exdir,paste0("Generic_",code,".xml")))
+
+    str_concepts <- xml_structure %>% xml_find_all("//structure:ConceptScheme")
+    var_concepts <- str_concepts %>%
+      xml_find_all("structure:Concept") %>%
+      xml_attr("id") %>% setdiff(c("TIME", "GEO", "OBS_VALUE", "OBS_STATUS" ))
+    concepts <- c("GEO",var_concepts)
+
+    concept_names <- lapply(concepts,function(c){
+      xml_structure %>% xml_find_all(paste0("//structure:ConceptScheme/structure:Concept[@id='",c,"']/structure:Name[@xml:lang='en']")) %>%
+        xml_text}) %>% unlist
+
+    concept_lookup <- rlang::set_names(concept_names,concepts)
+
+    descriptions_for_code <- function(c){
+      c <- gsub("0$","",c)
+      base <- xml_structure %>% xml_find_all(paste0("//structure:CodeList[@id='CL_",toupper(c),"']/structure:Code"))
+      desc_text = ifelse(length(xml_find_all(base[1] ,".//structure:Description[@xml:lang='en']"))==0,".//structure:Description",".//structure:Description[@xml:lang='en']")
+      rlang::set_names(
+        base %>% purrr::map(function(e){e %>% xml_find_all(desc_text) %>% xml_text}) %>% unlist %>% trimws(),
+        base %>% purrr::map(function(e){xml_attr(e,"value")}) %>% unlist
+      )
+    }
+
+    series <- xml_find_all(xml_data,"//generic:Series")
+
+    #series <- series[1:10]
+
+    #l <- lapply(concepts,function(c){series %>% xml_find_all(paste0(".//generic:Value[@concept='",c,"']")) %>% xml_attr("value")})
+
+    time_data <- function(series){
+      if (time_value) {
+        result=rep(time_value,length(series))
+      } else {
+        message("Extracting year data")
+        series %>% xml_find_all(".//generic:Time") %>% xml_text
+      }
+    }
+    value_data <- function(series){
+      message("Extracting values data")
+      #series %>% xml_find_all(".//generic:ObsValue") %>% xml_attr("value")
+      series %>% xml_find_all("./generic:Obs") %>% xml_find_first("./generic:ObsValue") %>% xml_attr("value")
+    }
+    code_data <- function(series,c){
+      message(paste0("Extracting data for ",concept_lookup[c]))
+      series %>% xml_find_all(paste0(".//generic:Value[@concept='",c,"']")) %>% xml_attr("value")
+    }
+
+    df=concepts %>%
+      purrr::map(function(c){code_data(series,c)}) %>%
+      rlang::set_names(concept_names) %>%
+      tibble::as.tibble()  %>%
+      dplyr::bind_cols(tibble::tibble(Year = time_data(series), Value = value_data(series)))
+
+    for (i in seq(1,length(concepts),1)) {
+      c=concepts[i]
+      n=concept_names[i] %>% as.name
+      lookup <- descriptions_for_code(c)
+      nid=paste0(n," ID") %>% as.name
+      df <- df %>%
+        mutate(!!nid := !!n) %>%
+        mutate(!!n := lookup[!!nid])
+    }
+
+    fix_ct_geo_format <- function(geo){
+      ifelse(nchar(geo)==9,paste0(substr(geo,1,7),".",substr(geo,8,9)),geo)
+    }
+
+    data <- df  %>%
+      rename(GeoUID=`Geography ID`) %>%
+      mutate(GeoUID = fix_ct_geo_format(GeoUID),
+             Value=as.numeric(Value))
+
+    unlink(exdir,recursive = TRUE)
+    saveRDS(data,path)
+  } else {
+    data <- readRDS(path)
+  }
+  data
+}
+
+#' Set up virtual conda env
+#' @export
+install_python_environment <- function(python_path="/anaconda3/bin/python3"){
+  reticulate::use_python(python_path)
+  reticulate::conda_binary(conda = python_path)
+  reticulate::conda_remove("r-reticulate")
+  reticulate::conda_create("r-reticulate",packages="lxml")
+}
+
+
+xml_to_csv <- function(code,url,python_path="/anaconda3/bin/python3",refresh=FALSE,time_value=NA,temp=NA){
+  path <- file.path(getOption("custom_data_path"),paste0(code,".csv"))
+  if (refresh | !file.exists(path)) {
+    remove_temp=FALSE
+    if (is.na(temp)) {
+      message("Downloading file")
+      remove_temp=TRUE
+      temp <- tempfile()
+      download.file(url,temp)
+    }
+    message("Converting xml to csv")
+    reticulate::use_python(python_path)
+    reticulate::conda_binary(conda = python_path)
+    reticulate::use_condaenv("r-reticulate")
+    statcan = reticulate::import_from_path("statcan",file.path(system.file(package="cancensusHelpers")))
+    statcan$convert_statcan_xml_to_csv(temp,path)
+    if (remove_temp) unlink(temp)
+    message("Done converting xml to csv")
+  }
+  path
+}
+
+
+#' Download xml xtab data from url, tag with code for caching
+#' useful for older (2006) data that does not come as csv option
+#'
+#' @param code The statcan table code, needed for caching and for locating the table to extract
+#' @param url The url to the xml table download. Only needed if the table is not cached, but
+#' reommended to always include for reproducibility
+#' @param python_path path to python executable, expects a conda environment 'r-reticulate' to exist.
+#' Can be set up with `install_python_environment`.
+#' @param refresh Will refresh cached data if set to TRUE, default is FALSE
+#' @param temp A path to the downloaded xml zip file. Useful if file has already been downloaded and should not be
+#' downloaded again.
+#'
+#' @export
+xml_via_python <- function(code,url,python_path="/anaconda3/bin/python3",refresh=FALSE,temp=NA){
+  readr::read_csv(xml_to_csv(code,url,python_path=python_path,refresh=refresh,temp=temp))
+}
+
+
 #' @export
 gather_for_grep_string <- function(data,gather_key,grep_string){
   vars <- names(data)[grepl(grep_string,names(data))]
   short_vars <- gsub(grep_string,"",vars)
   names(data) <- gsub(grep_string,"",names(data))
-  data %>% gather(key=!!gather_key,value="Value",short_vars)
+  data %>% dplyr::gather(key=!!gather_key,value="Value",short_vars)
 }
 
 #' @export
 strip_columns_for_grep_string <- function(data,grep_string){
-  data %>% select(names(data)[!grepl(grep_string,names(data))])
+  data %>% dplyr::select(names(data)[!grepl(grep_string,names(data))])
 }
 
 #' Simple key-value cache function accepting closures
@@ -341,7 +514,7 @@ get_cov_census_data <- function(year,use_cache=TRUE){
     row.names(dd)=dd$Variable
     d <- t(dd %>% select(-Variable))
     region_names <- rownames(d)
-    transposed_data <- as.tibble(d) %>%
+    transposed_data <- tibble::as.tibble(d) %>%
       dplyr::mutate_all(dplyr::funs(parse_number)) %>%
       mutate(NAME=case_when(
         grepl("CSD",region_names) ~ "City of Vancouver",
@@ -373,7 +546,7 @@ get_ecumene_2016 <- function(refresh=FALSE){
   read_sf(file.path(path,"lecu000e16a_e.shp"))
 }
 
-
+#' @import xml2
 #' @importFrom dplyr %>%
 #' @importFrom rlang .data
 NULL
